+++ CafeMaestro/RoastLogPage.xaml.cs
@@ -5,24 +5,25 @@ namespace CafeMaestro;
 public partial class RoastLogPage : ContentPage
 {
     private readonly RoastLogPageViewModel _viewModel;
+    private IDispatcherTimer? _ticker;
 
     public RoastLogPage(RoastLogPageViewModel viewModel)
     {
         InitializeComponent();
         BindingContext = _viewModel = viewModel;
-        _viewModel.AlertAsync = (title, message, cancel) => DisplayAlertAsync(title, message, cancel);
-        _viewModel.ActionSheetAsync = (title, cancel, destruction, buttons) =>
-            DisplayActionSheetAsync(title, cancel, destruction, buttons);
     }
 
     protected override async void OnAppearing()
     {
         base.OnAppearing();
         await _viewModel.OnAppearingAsync();
+        EnsureTicker();
+        _ticker!.Start();
     }
 
     protected override void OnDisappearing()
     {
+        _ticker?.Stop();
         _viewModel.OnDisappearing();
         base.OnDisappearing();
     }
@@ -31,5 +32,27 @@ public partial class RoastLogPage : ContentPage
     {
         _ = _viewModel.NavigateToRoastAsync();
         return true;
+    }
+
+    private void EnsureTicker()
+    {
+        if (_ticker is not null)
+        {
+            return;
+        }
+
+        _ticker = Dispatcher.CreateTimer();
+        _ticker.Interval = TimeSpan.FromSeconds(1);
+        _ticker.Tick += async (_, _) => await _viewModel.RefreshTimeProjectionAsync();
+    }
+
+    private void OnPageSizeChanged(object? sender, EventArgs e)
+    {
+        bool isWide = Width >= 600;
+        LogBody.ColumnDefinitions[1].Width = isWide
+            ? new GridLength(2, GridUnitType.Star)
+            : new GridLength(0);
+        DetailPane.IsVisible = isWide;
+        _viewModel.SetWideLayout(isWide);
     }
 }